First commit
To run it, you'll have to run the gradlew script (generated by gradle).
The IDE alternative is to set Intellij or eclipse to run the script. Basically,
it will download gradle (if it doesn't already exist on your machine), startup an
instance of tomcat (jvm application server) and launch the current code in a jar.

./gradlew build && java -jar build/libs/whitechapel-0.1.0.jar