Adding users and notifications
Users and notifications both required a bit of storage. I ended up using
H2 for the moment (and in memory database provided by spring) just to keep development
simple (no db setup required). When we deploy, all we need to do is swap out the adapter
for postgres or something and it'll work the same.

Committing the actual firebase keys for the moment (which is responsible for push notifications),
but we'll want to keep those out of source control in the future. They're currently configurable
through src/main/resources/notifications.properties, but always good to have an example
config file.